+++ src/lib.rs
@@ -9,14 +9,10 @@
 mod debounced;
 mod status;
 
-pub use debounced::Debounced;
+pub use debounced::{Debounced, DebouncedIntegrator, DebouncedShifter};
 pub use status::Status;
 pub mod active;
 pub mod strategy;
-
-/// Convenience for [`Debounced<_, Integrator, _>`]
-pub type DebouncedIntegrator<A, F> = Debounced<A, strategy::Integrator<A>, F>;
-//pub type DebouncedShift<A, F> = Debounced<A, strategy::Shift, F>;
 
 mod private {
 	pub trait Sealed {}